# master graph vertex & edge count speedup