Closes #200.
* Combined & tidied up some redundant calls to several functions that collectively did a bunch of redundant iterating through the HighwayGraph vertices.
  - **Python:** 25.6 -> 23.9 s, a 7.1% speedup
  - **C++:** 7.8 -> 6.8 s, a 14.7% speedup
* Moved the summary info printout of # of vertices & edges for each graph from before to after writing master graphs, to save either still computing it twice or needlessly passing a handful of extra variables to `write_master_graphs_tmg`. If we count this toward our master graph time savings:
  - **Python:** 26.5 -> 23.9 s, a 10.9% speedup
  - **C++:** 8.5 -> 6.8 s, a 25% speedup.