@yakra 2019-07-04
```
yakra@BiggaTomato:~/TravelMapping/yakra/DataProcessing/siteupdate/cplusplus$ cat siteupdate_m1.log | head -n 427 | tail -n 12
[66.8] Setting up subgraphs.
[66.9] Writing master TM graph files.
[67.0] Writing continent graphs.
AFR (22903,23405) (10646,11148) (10646,11148) ASI (49424,50335) (24413,25324) (24413,25324) OCE (11050,11391) (6452,6793) (6452,6793) SA (1491,1619) (612,740) (612,740) EUR (225299,258872) (166834,200407) (166838,200411) 
[72.1] Writing multisystem graphs.
USA-national (84015,89157) (71238,76380) (71244,76386) EUR-freeways (27252,27528) (19108,19384) (19109,19385) NA (258342,295570) (199642,236870) (199656,236884) 
[74.0] Writing system graphs.
usai (22399,22778) (18145,18524) (18148,18527) cantch (3440,3448) (2263,2271) (2263,2271)    Simple graph has 568486 vertices, 641192 edges.
Collapsed graph has 408576 vertices, 481282 edges.
 Traveled graph has 408594 vertices, 481300 edges.
usaus (61187,62788) (52550,54151) (52553,54154) 
[75.0] Writing country graphs.
```
OTOH, this looks a little bit ugly, and I'd like to find a nicer way to format it.
I might leave that 0.7s single-threaded, and pass that handful of variables to `write_master_graphs_tmg` after all...